requirements.txt
python-pptx
qrcode
openai
diffusers
transformers
accelerate
xformers
aiogram
python-dotenv